requirements/development.txt
apipkg==1.5
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
cookies==2.2.1
coreapi==2.3.3
coreschema==0.0.4
coverage==4.4.2
decorator==4.4.0
Django==2.0.1
django-cors-headers==1.2.0
django-extensions==1.9.9
django-filter==1.1.0
django-nose==1.4.5
django-postgres-copy==2.4.1
django-rest-swagger==2.2.0
django-test-without-migrations==0.6
djangorestframework==3.7.7
djangorestframework-gis==0.12
execnet==1.6.0
idna==2.6
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.0.0
networkx==2.1
nose==1.3.7
numpy==1.14.5
openapi-codec==1.3.2
packaging==16.8
pandas==0.22.0
pluggy==0.9.0
psycopg2==2.8.3
py==1.8.0
pyparsing==2.2.0
pytest==4.4.0
pytest-django==3.4.8
pytest-forked==1.0.2
pytest-xdist==1.28.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.18.4
responses==0.8.1
selenium==3.141.0
simplejson==3.13.2
six==1.11.0
tblib==1.3.2
typing==3.6.6
uritemplate==3.0.0
urllib3==1.22
whitenoise==4.0b4
